FR-9637-support-server-components-for-next-13
