+++ apps/demo-saas/app/layout.tsx
@@ -1,0 +1,12 @@
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html>
+      <head></head>
+      <body>{children}</body>
+    </html>
+  );
+}
