+++ apps/demo-saas/next.config.js
@@ -10,6 +10,9 @@ const nextConfig = {
     // See: https://github.com/gregberge/svgr
     svgr: false,
   },
+  experimental:{
+    appDir: true
+  },
 };
 
 module.exports = withNx(nextConfig);
