+++ libs/nextjs/common/FronteggConfig.ts
