+++ libs/nextjs/common/helpers.ts
@@ -1,15 +1,12 @@
-import { ServerResponse } from 'http';
 import cookie, { CookieSerializeOptions } from 'cookie';
+import { ServerResponse } from 'http';
 import { sealData } from 'iron-session';
-import fronteggConfig from './FronteggConfig';
 import { decodeJwt } from 'jose';
-import { NextApiRequest, NextPageContext } from 'next/dist/shared/lib/utils';
-import { FronteggNextJSSession } from './types';
-import { fronteggRefreshTokenUrl, fronteggSilentRefreshTokenUrl } from '@frontegg/rest-api';
-import { getHostedLoginRefreshToken, getSession } from './session';
 import * as zlib from 'zlib';
+import fronteggConfig from '../common/FronteggConfig';
+import { CreateCookieArguments } from '../common/types';
 
-function rewriteCookieProperty(
+export function rewriteCookieProperty(
   header: string | string[],
   config: any,
   property: string
@@ -42,166 +39,7 @@ function rewriteCookieProperty(
   );
 }
 
-async function refreshTokenHostedLogin(
-  ctx: NextPageContext,
-  headers: Record<string, string>
-): Promise<Response | null> {
-
-  const refreshToken = await getHostedLoginRefreshToken(ctx.req!);
-
-  if (!refreshToken) {
-    return null;
-  }
-
-  return await fetch(
-    `${process.env['FRONTEGG_BASE_URL']}/frontegg${fronteggSilentRefreshTokenUrl}`,
-    {
-      method: 'POST',
-      credentials: 'include',
-      body: JSON.stringify({
-        grant_type: 'refresh_token',
-        refresh_token: refreshToken,
-      }),
-      headers: {
-        'accept-encoding': headers['accept-encoding'],
-        'accept-language': headers['accept-language'],
-        cookie: headers['cookie'],
-        accept: headers['accept'],
-        'user-agent': headers['user-agent'],
-        connection: headers['connection'],
-        'cache-control': headers['cache-control'],
-      },
-    }
-  );
-
-}
-
-async function refreshTokenEmbedded(
-  ctx: NextPageContext,
-  headers: Record<string, string>,
-  cookies: Record<string, any>,
-): Promise<Response | null> {
-
-  const refreshTokenKey = `fe_refresh_${fronteggConfig.clientId}`.replace(
-    /-/g,
-    ''
-  );
-  const cookieKey = Object.keys(cookies).find((cookie) => {
-    return cookie.replace(/-/g, '') === refreshTokenKey;
-  });
-
-  if (!cookieKey) {
-    // ctx.res?.setHeader('set-cookie', removedCookies);
-    // remove all fe_nextjs-session cookies
-    return null;
-  }
-
-  return await fetch(
-    `${process.env['FRONTEGG_BASE_URL']}/frontegg${fronteggRefreshTokenUrl}`,
-    {
-      method: 'POST',
-      credentials: 'include',
-      body: '{}',
-      headers: {
-        'accept-encoding': headers['accept-encoding'],
-        'accept-language': headers['accept-language'],
-        cookie: headers['cookie'],
-        accept: headers['accept'],
-        'user-agent': headers['user-agent'],
-        connection: headers['connection'],
-        'cache-control': headers['cache-control'],
-      },
-    }
-  );
-}
-
-export async function refreshToken(
-  ctx: NextPageContext
-): Promise<FronteggNextJSSession | null> {
-  try {
-    const request = ctx.req;
-    if (!request) {
-      return null;
-    }
-    try {
-      const session = await getSession(ctx.req as any);
-      if (session) {
-        return session;
-      }
-    } catch (e) {
-      // ignore error catch
-    }
-    const isSecured = new URL(fronteggConfig.appUrl).protocol === 'https:';
-    const headers = request.headers as Record<string, string>;
-    const cookies = (request as NextApiRequest).cookies;
-
-    if (ctx.req!.url!.startsWith('/oauth/callback')) {
-      return null;
-    }
-    let response: Response | null;
-    if (fronteggConfig.fronteggAppOptions.hostedLoginBox) {
-      response = await refreshTokenHostedLogin(ctx, headers);
-    } else {
-      response = await refreshTokenEmbedded(ctx, headers, cookies);
-    }
-    if (!response) {
-      removeCookies(
-        fronteggConfig.cookieName,
-        isSecured,
-        fronteggConfig.cookieDomain,
-        ctx.res!
-      );
-      return null;
-    }
-
-    if (response.ok) {
-      const data = await response.text();
-      const rewriteCookieDomainConfig = {
-        [fronteggConfig.baseUrlHost]: fronteggConfig.cookieDomain,
-      };
-      // @ts-ignore
-      const cookieHeader = response.headers.raw()['set-cookie'];
-      let newSetCookie = rewriteCookieProperty(
-        cookieHeader,
-        rewriteCookieDomainConfig,
-        'domain'
-      );
-      const [ session, decodedJwt, refreshToken ] = await createSessionFromAccessToken(data);
-      if (!session) {
-        return null;
-      }
-      const cookieValue = createCookie({ session, expires: new Date(decodedJwt.exp * 1000), isSecured })
-      if (typeof newSetCookie === 'string') {
-        newSetCookie = [ newSetCookie ];
-      }
-      newSetCookie.push(...cookieValue);
-      ctx.res?.setHeader('set-cookie', newSetCookie);
-      return {
-        accessToken: JSON.parse(data).accessToken,
-        user: decodedJwt,
-        refreshToken
-      };
-    } else {
-      // remove all fe_nextjs-session cookies
-      // ctx.res?.setHeader('set-cookie', removedCookies);
-      return null;
-    }
-  } catch (e) {
-    return null;
-  }
-}
-
-type CreateCookieArguments = {
-  cookieName?: string,
-  session: string,
-  expires: CookieSerializeOptions['expires'],
-  isSecured: CookieSerializeOptions['secure'],
-  cookieDomain?: CookieSerializeOptions['domain'],
-  httpOnly?: CookieSerializeOptions['httpOnly'],
-  path?: CookieSerializeOptions['path']
-}
 const COOKIE_MAX_LENGTH = 4096
-
 export function createCookie(
   {
     cookieName = fronteggConfig.cookieName,
@@ -222,7 +60,7 @@ export function createCookie(
   }
   const cookieValue = cookie.serialize(cookieName, session, options);
   if (cookieValue.length < COOKIE_MAX_LENGTH) {
-    return [ cookieValue ]
+    return [cookieValue]
   }
   return createSplitCookie(cookieName, session, options, cookieValue.length)
 }
@@ -265,9 +103,9 @@ export function addToCookies(newCookies: string[], res: ServerResponse) {
   let existingSetCookie =
     (res.getHeader('set-cookie') as string[] | string) ?? [];
   if (typeof existingSetCookie === 'string') {
-    existingSetCookie = [ existingSetCookie ];
+    existingSetCookie = [existingSetCookie];
   }
-  res.setHeader('set-cookie', [ ...existingSetCookie, ...newCookies ]);
+  res.setHeader('set-cookie', [...existingSetCookie, ...newCookies]);
 }
 
 export function removeCookies(
@@ -280,9 +118,9 @@ export function removeCookies(
   let existingSetCookie =
     (res.getHeader('set-cookie') as string[] | string) ?? [];
   if (typeof existingSetCookie === 'string') {
-    existingSetCookie = [ existingSetCookie ];
+    existingSetCookie = [existingSetCookie];
   }
-  res.setHeader('set-cookie', [ ...existingSetCookie, ...cookieValue ]);
+  res.setHeader('set-cookie', [...existingSetCookie, ...cookieValue]);
 }
 
 export function compress(input: string): Promise<string> {
@@ -311,7 +149,7 @@ export function uncompress(input: string): Promise<string> {
 
 export async function createSessionFromAccessToken(
   output: string
-): Promise<[ string, any, string ] | []> {
+): Promise<[string, any, string] | []> {
   try {
     const data = JSON.parse(output);
     const accessToken = data?.accessToken ?? data.access_token;
@@ -327,7 +165,7 @@ export async function createSessionFromAccessToken(
       password: fronteggConfig.passwordsAsMap,
       ttl: decodedJwt.exp,
     });
-    return [ session, decodedJwt, refreshToken ];
+    return [session, decodedJwt, refreshToken];
   } catch (e) {
     return [];
   }
