+++ package.json
@@ -15,23 +15,23 @@
   ],
   "private": true,
   "dependencies": {
-    "@nrwl/next": "14.0.5",
+    "@nrwl/next": "^15.0.10",
     "core-js": "^3.6.5",
-    "next": "^12.2.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "next": "^13.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.3.0"
   },
   "devDependencies": {
-    "@nrwl/cli": "14.0.5",
-    "@nrwl/cypress": "14.0.5",
-    "@nrwl/eslint-plugin-nx": "14.0.5",
-    "@nrwl/jest": "14.0.5",
-    "@nrwl/linter": "14.0.5",
-    "@nrwl/react": "14.0.5",
-    "@nrwl/web": "14.0.5",
-    "@nrwl/workspace": "14.0.5",
+    "@nrwl/cli": "^15.0.10",
+    "@nrwl/cypress": "^15.0.10",
+    "@nrwl/eslint-plugin-nx": "^15.0.10",
+    "@nrwl/jest": "^15.0.10",
+    "@nrwl/linter": "^15.0.10",
+    "@nrwl/react": "^15.0.10",
+    "@nrwl/web": "^15.0.10",
+    "@nrwl/workspace": "^15.0.10",
     "@swc/core": "^1.2.177",
     "@swc/jest": "0.2.20",
     "@testing-library/react": "13.1.1",
@@ -42,9 +42,10 @@
     "@typescript-eslint/eslint-plugin": "~5.18.0",
     "@typescript-eslint/parser": "~5.18.0",
     "babel-jest": "27.5.1",
+    "css-loader": "^6.7.1",
     "cypress": "^9.1.0",
     "eslint": "~8.12.0",
-    "eslint-config-next": "12.1.5",
+    "eslint-config-next": "13.0.2",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "eslint-plugin-import": "2.26.0",
@@ -52,7 +53,7 @@
     "eslint-plugin-react": "7.29.4",
     "eslint-plugin-react-hooks": "4.4.0",
     "jest": "27.5.1",
-    "nx": "14.0.5",
+    "nx": "^15.0.10",
     "prettier": "^2.5.1",
     "react-test-renderer": "18.0.0",
     "ts-jest": "27.1.4",
